Feature : Create Aggregate with root is completed

+++ Sample/1.Core/MiniBlog.Core.Domain/Advertisements/DomainEvents/AdvertisementUnPublished.cs
@@ -2,4 +2,4 @@
 
 namespace MiniBlog.Core.Domain.Advertisements.DomainEvents;
 
-public sealed record AdvertisementUnPublished(Guid BusinessId, int Id) : IDomainEvent;
+public sealed record AdvertisementUnPublished(Guid BusinessId, long Id) : IDomainEvent;
